backend/requirements.txt
flask
flask-cors
google-generativeai>=0.3.1
